ota_handler.py, client.py: add offline install support for OTA
Resolves #197
Update the client and ota_handler classes to work in offline mode,
where offline mode means do not talk to the cloud.  Added a script
that uses the existing ota_handler class.  The ota package is expected
to be in the runtime/download directory.
Added a README with example details.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>