validate_scripts: update validate scripts to test alarm republish
Resolves #206
Update validate_app to publish 5 alarms with republish=True.  Update
validate_script to check for all 5.  In this test, republish is True,
which means all alarms will be republished.  If republish=False, then
only one alarm would be published since all parameters are the same.

Cleanup the validate_script to remove the query for app tokens.  This
required admin access.  Update the prompt to ask for a token.  Also,
add the token to the .travis.yml encrypted var list.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>