relay, device_manager: add support for reconnecting local tcp socket.
Resolves #228
When executing a remote access session to a device requesting access
to a web server, many web servers drop the local socket connection
shortly after the initiall connect and serve of data.  Add a reconnect
flag to the work flow so that the relay class will reconnect when the
local socket to the service drops.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>