relay.py: rewrite the relay.py using the high level interface.
Resolves #241
The low level interface did not select on a wsock consistantly.  A
work around was added to force a read on the websocket with the caveat
that it could block on the read.  This commit removes the low level
implementation and the work around.  This relay uses the WebSocketApp
class.

Updated unittest to adapt to the changes.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>