ota_handler.py: handle return code for agent update
Resolves #249
If the agent is reset or updated, currently the transaction ID is left
to timeout.  Write the transaction ID and status to a file and read that
file when the device manager starts.  Then, during initialization, check
for the transaction file and ack the actions.  The transaction file
supports multiple entries of the format:
	<ID>,<STATUS>
Currently, only the device manager and the ota handler use this
functionality and only for agent reset and software update.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>