handler.py, tr50.py: batch telemetry, attributes, location, alarms
Resolves #256

The agent loads telemetry, attributes, location and alarm data into a
msg queue and then processes the queue during each iteration.  The data is
collected in a batch manner but the *.batch cloud APIs were not used.
This change updates the telemetry, attributes, location and alarm APIs
to use the batch cloud APIs.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>