client.py, handler.py: add idle sleep to the TX path
Resolves # 256
RCA:
There is an API/s limit per ORG enforced by the cloud.  All producer
operations on the device invoke a cloud API to publish telemetry,
location, attributes and alarms.   The device cloud API TX path can
flood the cloud with API bursts that return error -90031.  This is an
important error and depending on the scenario, this means:
  * API/s threshold for the ORG has been crossed.
  * Any data involved in the transaction will be dropped by the cloud
    when this error occurs.
  * The application runs the risk of being banned by the cloud.

The TX flood can be caused by two scenarios:
  1) when connected, it is possible to call more API/s calls than allowed.
  2) when disconnected, it is possible for the Paho driver to queue up
     more API/s calls than are allowed.  Once reconnected, this Paho
     queue will flood the cloud.

Solution 1) is addressed by the new batch APIs and by adding a small
delay on sending data.  The default idle sleep is 0.1 seconds.  This
is tunable in the application if the device is flooding the cloud:

```
client.idle_sleep = 0.2
```
This commit also adds support for an error handler to be provided by
the application which will be called when a cloud error is received.
Note: you may receive several notifications depending on how fast you
are sending data.  E.g.: the application registers an optional error
handler when initializing the class:
```
def my_error_handler(err_list, sent_msg, reply):
   global system_error
   print("System Error: error: {} sent: {} reply: {}".format(err_list, sent_msg, reply))
   system_error = True

(in main function)
client = iot.Client(app_id, error_handler=my_error_handler)
```
When the applications error handler will be invoked as soon as the
session receives a cloud error and the application has a chance to
react.

Solution 2) will require changes to the Paho module, which is outside
the scope of this commit.  For the most part, the new batch APIs will
mitigate the problem.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>