admin_tools: update device role permissions.
Resolves #266
The device role permissions were not strict enough and allowed many
"delete" operations.  As a default role example for the device, it
makes sense to limit the API access to what is required for security
reasons.

Note: HDC does not control the role definition; HDC
might initially populate an ORG with roles for a convenience, but
someone with admin privelige must maintain the roles, thing defs etc.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>