_core: add support for the corr ID for telemetry
Resolves #274
A correlation ID can be sent in a telemetry sample so that it can be
read back and verified.  However, this functionality was added only in
cloud version 18.05.  The currently cloud version is 18.04.03 which
does not return the correlation ID when queried.  This commit will
enable the corr_id field when the cloud is updated to support it.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>