device_manager.py: add callback to refresh the remote access ports
Resolves#275
In order to support dynamic services, e.g. vncserver, we need to be
able to refresh the remote access attribute with an updated list of
ports.  This commit adds a callback that is called on device manager
init and can be invoked any time, e.g. via rest.  When called the
remote_access_support attribute is update with the ports defined in
iot.cfg that are actually listening.  The callback also returns an
out parameter json string with the same data so that it can be called
from REST.

Add iot.cfg variable to disable ports scanned on startup of the device
manager.  The default is to scan service ports.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>