ota_handler: add support for passing parameters into install scripts
Resolves #285

Add a parameter that is defined in the thing def that specifies "extra
parameters".  Pass the extra parameters to the syscall that runs the
install phase scripts.  The extra parameters will be in the environal
variable "HDC_EXTRA_PARAMS".  This will allow one ota pkg to be
deployed via triggers with extra parameters for many different
installations.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>