client.py: add support for a Null app_id
Resolves #278
The thing key is made up of the device ID + app ID.  Some use cases
may not want the app_id, e.g. migration from previous clouds with
historical data.  In this case, it is suitable to specify None as the
app_id.  Note: the caveat is that only one app can use the device ID,
that is why we append the app_id.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>